de-12: rename mocks folder for graph flow
